extract to conftest file
